+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # Sable Client Changelog
 
-## 1.3.0 - in dev
+## 1.3.0 - 3/2/2026
 
 - Mobile push notifications! ([#44](https://github.com/7w1/sable/issues/44), [#49](https://github.com/7w1/sable/pull/49)) - [Evie Gauthier](https://github.com/Just-Insane)
-- Beta Simplified Sliding Sync support ([#67](https://github.com/7w1/sable/pull/67)) - [Evie Gauthier](https://github.com/Just-Insane)
+- Beta Simplified Sliding Sync support ([#67](https://github.com/7w1/sable/pull/67), ([#75](https://github.com/7w1/sable/pull/75)) - [Evie Gauthier](https://github.com/Just-Insane)
 - Codebase cleanups, CI improvements, and docker builds ([#26](https://github.com/7w1/sable/pull/26), [#35](https://github.com/7w1/sable/pull/35), [#62](https://github.com/7w1/sable/pull/62), [#64](https://github.com/7w1/sable/pull/64), [#65](https://github.com/7w1/sable/pull/65)) - [haz](https://github.com/hazre)
+- Add room/space specific pronouns, when enabled by room/space admin. ([#30](https://github.com/7w1/sable/issues/30))
 - Add validation to timezones before rendering.
 - Fix invalid matrix.to event link generation ([cinnyapp#2717](https://github.com/cinnyapp/cinny/pull/2717)) - [tulir](https://github.com/tulir)
 - Fix Call Rooms' chat button ([#58](https://github.com/7w1/sable/pull/58)) - [Rosy-iso](https://github.com/Rosy-iso)